backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.5
yfinance==0.2.36
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.1
authlib==1.3.0
gunicorn==21.2.0
eventlet==0.35.2
